requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
click==8.1.7
MarkupSafe==2.1.5
python-dotenv==1.2.1
psycopg[binary]==3.2.12
